add option to run build commands in bwrap, optionally without network access
requires bwrap to be installed of course.

bwrap is used to limit the system view, only to /bin /lib /usr, and the ding
build, home and toolchain dirs.

if net isolation is enabled, the build script doesn't have a network device.
just loopback. so it can't make outside connections.

also fix bug where we didn't prefix the build script with the commands from
config.Run.